+++ order/domain/servicePorts.go
@@ -3,14 +3,15 @@ package domain
 import (
 	"context"
 
-	"flamingo.me/flamingo/v3/core/oauth/domain"
+	"flamingo.me/flamingo/v3/core/auth"
 )
 
 type (
-
-	// CustomerOrderService interface
-	CustomerOrderService interface {
-		Get(context.Context, domain.Auth) ([]*Order, error)
-		GetByID(context.Context, domain.Auth, string) (*Order, error)
+	// CustomerIdentityOrderService loads orders for an authenticated user
+	CustomerIdentityOrderService interface {
+		// Get all orders for a customer
+		Get(ctx context.Context, identity auth.Identity) ([]*Order, error)
+		// GetByID returns a single order for a customer
+		GetByID(ctx context.Context, identity auth.Identity, orderID string) (*Order, error)
 	}
 )